server/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
google-genai
httpx
prisma
pydantic
python-multipart
